Rewrite the compiler.
The primary change here is the removal of the 'Jump' instruction.
Removing this instruction should in theory make execution slightly
faster overall, since there will be fewer instructions to execute.
The 'Jump' instruction is replaced by an explicit 'goto' field on
all of the other instructions (sans 'Match' and 'Split').

The impetus for this change is as a precursor to switching to a byte
based automaton. In particular, in order to efficiently compile
a UTF-8 automaton, we must factor out common suffixes in the byte-based
character classes, or else suffer dramatically increased program sizes.
Factoring out common suffixes does lead to a more complicated program
however, which would need to heavily rely on the 'Jump' instruction if
it existed, which doesn't seem optimal to me (and would negate some of
the size benefit of removing common suffixes).

In the process of rewriting the compiler, I've also attempted to make
the setting of the 'goto' fields a bit more clearer with an explicit
representation for "Holes." It doesn't encode as many invariants into the
type system as I'd like, but it should at least prevent errant hole-filling
at runtime (panic'ing on an invariant violation).